Move add_resident_activity to helper
